+++ src/TestEngine/nunit.engine.core/Utilities/ResultHelper.cs
@@ -1,4 +1,4 @@
-﻿// ***********************************************************************
+// ***********************************************************************
 // Copyright (c) 2015 Charlie Poole, Rob Prouse
 //
 // Permission is hereby granted, free of charge, to any person obtaining
@@ -27,7 +27,7 @@ using System.Globalization;
 using System.Reflection;
 using System.Xml;
 
-namespace NUnit.Engine.Internal
+namespace NUnit.Engine.Utilities
 {
     /// <summary>
     /// ResultHelper provides static methods for working with
