+++ src/TestEngine/nunit.engine.tests/Services/ResultWriters/XmlOutputTest.cs
@@ -1,4 +1,4 @@
-﻿// ***********************************************************************
+// ***********************************************************************
 // Copyright (c) 2011 Charlie Poole, Rob Prouse
 //
 // Permission is hereby granted, free of charge, to any person obtaining
@@ -26,6 +26,7 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Reflection;
+using NUnit.Engine.Utilities;
 
 namespace NUnit.Engine.Services.ResultWriters.Tests
 {
