+++ src/TestEngine/nunit.engine.tests/Utilities/ResultHelperTests.cs
@@ -1,4 +1,4 @@
-﻿// ***********************************************************************
+// ***********************************************************************
 // Copyright (c) 2015 Charlie Poole, Rob Prouse
 //
 // Permission is hereby granted, free of charge, to any person obtaining
@@ -23,7 +23,7 @@
 
 using System.Xml;
 
-namespace NUnit.Engine.Internal.Tests
+namespace NUnit.Engine.Utilities
 {
     using Framework;
