+++ src/TestEngine/nunit.engine/Runners/MasterTestRunner.cs
@@ -1,4 +1,4 @@
-﻿// ***********************************************************************
+// ***********************************************************************
 // Copyright (c) 2011-2014 Charlie Poole, Rob Prouse
 //
 // Permission is hereby granted, free of charge, to any person obtaining
@@ -22,6 +22,7 @@
 // ***********************************************************************
 
 using System;
+using System.ComponentModel;
 using System.Diagnostics;
 using System.Globalization;
 using System.IO;
@@ -30,7 +31,7 @@ using System.Xml;
 using NUnit.Common;
 using NUnit.Engine.Internal;
 using NUnit.Engine.Services;
-using System.ComponentModel;
+using NUnit.Engine.Utilities;
 
 namespace NUnit.Engine.Runners
 {
